move components out to own functions
